move the 'IS_PLUGIN_INIT_COMPLETED' field to 'PluginFinder'
