+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/conf/SnifferConfigInitializer.java
@@ -50,7 +50,6 @@ public class SnifferConfigInitializer {
     private static final String ENV_KEY_PREFIX = "skywalking.";
     private static Properties AGENT_SETTINGS;
     private static boolean IS_INIT_COMPLETED = false;
-    private static boolean IS_PLUGIN_INIT_COMPLETED = false;
 
     /**
      * If the specified agent config path is set, the agent will try to locate the specified agent config. If the
@@ -234,13 +233,5 @@ public class SnifferConfigInitializer {
             default:
                 LogManager.setLogResolver(new PatternLogResolver());
         }
-    }
-
-    public static void pluginInitCompleted() {
-        IS_PLUGIN_INIT_COMPLETED = true;
-    }
-
-    public static boolean isPluginInitCompleted() {
-        return IS_PLUGIN_INIT_COMPLETED;
     }
 }
